+++ activejob/lib/active_job/logging.rb
@@ -47,7 +47,13 @@ module ActiveJob
       end
 
       def logger_tagged_by_active_job?
-        logger.formatter.current_tags.include?("ActiveJob")
+        if logger.formatter.respond_to?(:current_tags)
+          logger.formatter.current_tags.include?("ActiveJob")
+        else
+          # Return true because if we cannot figure out what the current tags
+          # are, we do not want to keep adding the ActiveJob tag
+          true
+        end
       end
 
     class LogSubscriber < ActiveSupport::LogSubscriber #:nodoc: